% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDataFilter_df.R
\name{getInitializationCode.shinyDataFilter_df}
\alias{getInitializationCode.shinyDataFilter_df}
\title{Handler for retrieving static code for a shinyDataFilter_df}
\usage{
getInitializationCode.shinyDataFilter_df(obj, name = NULL)
}
\arguments{
\item{obj}{shinyDataFilter_df object}

\item{name}{unused}
}
\value{
Returns the 'code' attribute as a string from any object of class
  shinyDataFilter_df
}
\description{
A function for scriptgloss::getInitializationCode to dispatch to
}
\keyword{internal}
