% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDEAFilter_item.R
\name{IDEAFilter_item_ui}
\alias{IDEAFilter_item_ui}
\title{A single filter item as part of a IDEA filter module panel}
\usage{
IDEAFilter_item_ui(id)
}
\arguments{
\item{id}{a module id name}
}
\value{
a shiny \code{\link[shiny]{wellPanel}} to house the filter
}
\description{
This is a wrapper for \code{\link{shiny_data_filter_item_ui}} created to
match up with the module server function \code{\link{IDEAFilter_item}}.
}
\keyword{internal}
