% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.empty}
\alias{is.empty}
\title{Check if string is blank ""}
\usage{
is.empty(x)
}
\arguments{
\item{x}{text to check if equal to ""}
}
\value{
logical of blank or not
}
\description{
Check if string is blank ""
}
