% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullor}
\alias{nullor}
\alias{\%||\%}
\title{Shorthand operator for NULL fallback}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{left hand side for null-or operation}

\item{rhs}{right hand side for null-or operation}
}
\description{
Shorthand operator for NULL fallback
}
\keyword{internal}
