% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_vector_filter_numeric_few.R
\name{shiny_vector_filter_numeric_few}
\alias{shiny_vector_filter_numeric_few}
\title{A vector filter for numeric variables with only a few choices}
\usage{
shiny_vector_filter_numeric_few(
  input,
  output,
  session,
  x = shiny::reactive(factor()),
  filter_na = shiny::reactive(FALSE),
  verbose = FALSE
)
}
\description{
A vector filter for numeric variables with only a few choices
}
