% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_vector_filter_numeric_many.R
\name{shiny_vector_filter_numeric_many}
\alias{shiny_vector_filter_numeric_many}
\title{A vector filter for numeric variables with only many choices}
\usage{
shiny_vector_filter_numeric_many(
  input,
  output,
  session,
  x = shiny::reactive(numeric()),
  filter_na = shiny::reactive(FALSE),
  verbose = FALSE
)
}
\description{
A vector filter for numeric variables with only many choices
}
