% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_vector_filter_numeric_many.R
\name{shiny_vector_filter_numeric_many}
\alias{shiny_vector_filter_numeric_many}
\title{A vector filter for numeric variables with only many choices}
\usage{
shiny_vector_filter_numeric_many(
  input,
  output,
  session,
  x = shiny::reactive(numeric()),
  filter_na = shiny::reactive(FALSE),
  filter_expr = NULL,
  verbose = FALSE
)
}
\arguments{
\item{input}{requisite shiny module field specifying incoming ui input
reactiveValues}

\item{output}{requisite shiny module field capturing output for the shiny
data filter ui}

\item{session}{requisite shiny module field containing the active shiny
session}

\item{x}{The TODO}

\item{filter_na}{The \code{logical} TODO}

\item{filter_expr}{A character string that can specify initial filtering}

\item{verbose}{a \code{logical} value indicating whether or not to print log
statements out to the console}
}
\value{
a \code{\link[shiny]{reactiveValues}} list containing a logical
  vector called "mask" which can be used to filter the provided vector and an
  element "code" which is the expression used to generate the mask.
}
\description{
A vector filter for numeric variables with only many choices
}
\keyword{internal}
